LayoutTests/platform/qt/plugins/document-open-expected.txt

CONSOLE MESSAGE: PLUGIN: DOCUMENT OPEN SUCCESS
ERROR: nil result from [documentElement innerText]
